pyproject.toml
[project]
name = "tako"
version = "0.1.0"
description = "Tako chess engine with Hierarchical Reasoning Model (HRM)"
authors = [{name = "zfdupont"}]
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "einops>=0.7.0",
    "ray[default]>=2.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "maturin>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["games", "model"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

# Note: Maturin build system will be added in Phase 2+ when Rust code is introduced
# [tool.maturin]
# python-source = "python"
# module-name = "tako._bindings"
